+++ includes/vec4/vec4.h
@@ -13,8 +13,8 @@
 
 /* ******************************************************************************** */
 
-struct vec4 directionVec(double, double, double);
-struct vec4 pointVec(double, double, double);
+struct vec4 directionVec(f32, f32, f32);
+struct vec4 pointVec(f32, f32, f32);
 
 /* ******************************************************************************** */
 
@@ -23,23 +23,23 @@ struct vec4 subVec(struct vec4, struct vec4);
 
 /* ******************************************************************************** */
 
-double magnitude(struct vec4);
+f32 magnitude(struct vec4);
 struct vec4 normalize(struct vec4);
 struct vec4 scalarMul(struct vec4, int);
 
 /* ******************************************************************************** */
 
 struct vec4 componentProduct(struct vec4, struct vec4);
-double dotProduct(struct vec4, struct vec4);
+f32 dotProduct(struct vec4, struct vec4);
 struct vec4 crossProduct(struct vec4, struct vec4);
 
 /* ******************************************************************************** */
 
 struct vec4 {
-    double x;
-    double y;
-    double z;
-    double w;
+    f32 x;
+    f32 y;
+    f32 z;
+    f32 w;
 };
 
 /* ******************************************************************************** */
